Move feature functions to separate helpers file
