+++ pages/blink_handler.py
@@ -23,6 +23,7 @@ import os
 from framework import basehandlers
 from framework import permissions
 from internals import core_models
+from internals import feature_helpers
 from internals import user_models
 import settings
 from api.channels_api import construct_chrome_channels_details
@@ -111,7 +112,7 @@ class SubscribersHandler(basehandlers.FlaskHandler):
   def get_template_data(self, **kwargs):
     users = user_models.FeatureOwner.query().order(
         user_models.FeatureOwner.name).fetch(None)
-    feature_list = core_models.Feature.get_chronological()
+    feature_list = feature_helpers.get_chronological()
 
     milestone = self.get_int_arg('milestone')
     if milestone is not None:
